Componentized base stories.
