+++ components/00-base/layouts/grid/grids.stories.js
@@ -1,0 +1,23 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+import grid from './grid.twig'
+
+const Grid = (
+  grid({
+    grid_label: "Default",
+    items: {
+      1: "",
+      2: "",
+      3: ""
+    }
+  })
+)
+
+/**
+ * Add storybook definition for Grids.
+ */
+storiesOf('Base/Grids', module)
+  .add('Default', () => (
+    <div dangerouslySetInnerHTML={{__html: Grid }}></div>
+  ))
